style: Fish 엔티티 필드 추가 및 Point 타입변환 #101
Fish: fishInfo(물고기 설명) 필드 추가
Point: Long에서 int로 타입변환

+++ src/main/java/com/finalproject/seatudy/infra/GoogleMemberService.java
@@ -49,7 +49,7 @@ public class GoogleMemberService {
         String googleAC = jwtTokenUtils.generateJwtToken(googleMember);
         memberService.tokenToHeaders(googleAC, response);
 
-        Long point = memberService.calculateCurrentPoint(googleMember);
+        int point = memberService.calculateCurrentPoint(googleMember);
         log.info("Google 로그인 완료: {}",googleMember.getEmail());
         return ResponseDto.success(MemberResDto.fromEntity(googleMember, point));
     }
